refactor: extract general/automations/secrets sections
Move three settings sections out of the 852-LoC general.tsx route file
into per-section components under components/settings/. SectionHeader
moves to a shared helpers file. useTeam is hoisted to the route file
and passed as a prop so the shared query isn't fetched twice.
