+++ packages/web/src/routes/teams/$teamId/settings/general.tsx
@@ -2,6 +2,10 @@ import { createFileRoute } from '@tanstack/react-router';
 import { Copy, ExternalLink, Loader2, Plus, RefreshCw, Trash2 } from 'lucide-react';
 import { useEffect, useState } from 'react';
 import { RevisionsPanel } from '../../../../components/revisions-panel';
+import { AutomationsSection } from '../../../../components/settings/automations-section';
+import { GeneralSection } from '../../../../components/settings/general-section';
+import { SectionHeader } from '../../../../components/settings/helpers';
+import { SecretsSection } from '../../../../components/settings/secrets-section';
 import { Badge } from '../../../../components/ui/badge';
 import { Button } from '../../../../components/ui/button';
 import { Input } from '../../../../components/ui/input';
@@ -19,14 +23,13 @@ import {
 	useRestorePreferenceRevision,
 	useUpdatePreferences,
 } from '../../../../hooks/use-preferences';
-import { useCreateSecret, useDeleteSecret, useSecrets } from '../../../../hooks/use-secrets';
 import {
 	useCreateSkill,
 	useDeleteSkill,
 	useSkills,
 	useSyncSkill,
 } from '../../../../hooks/use-skills';
-import { DEFAULT_SUBTASK_PAGE_SIZE, useTeam, useUpdateTeam } from '../../../../hooks/use-teams';
+import { useTeam } from '../../../../hooks/use-teams';
 
 const settingsNav = [
 	{ id: 'general', label: 'General' },
@@ -42,6 +45,7 @@ const settingsNav = [
 
 function SettingsPage() {
 	const { teamId } = Route.useParams();
+	const { data: team } = useTeam(teamId);
 	const [activeSection, setActiveSection] = useState('general');
 
 	function scrollTo(id: string) {
@@ -70,10 +74,10 @@ function SettingsPage() {
 
 			<div className="space-y-8">
 				<div id="settings-general">
-					<GeneralSection teamId={teamId} />
+					<GeneralSection team={team} />
 				</div>
 				<div id="settings-automations">
-					<AutomationsSection teamId={teamId} />
+					<AutomationsSection teamId={teamId} team={team} />
 				</div>
 				<div id="settings-secrets">
 					<SecretsSection teamId={teamId} />
@@ -98,246 +102,6 @@ function SettingsPage() {
 				</div>
 			</div>
 		</div>
-	);
-}
-
-function SectionHeader({ title, desc }: { title: string; desc?: string }) {
-	return (
-		<div className="mb-4">
-			<h2 className="text-base font-medium">{title}</h2>
-			{desc && <p className="text-[13px] text-text-muted mt-1">{desc}</p>}
-		</div>
-	);
-}
-
-function GeneralSection({ teamId }: { teamId: string }) {
-	const { data: team } = useTeam(teamId);
-	return (
-		<section>
-			<SectionHeader title="General" desc="Basic team information." />
-			<div className="space-y-3 max-w-md">
-				<div>
-					<span className="text-xs font-medium uppercase tracking-wider text-text-muted block mb-1.5">
-						Team name
-					</span>
-					<div className="text-[13px]">{team?.name ?? '—'}</div>
-				</div>
-				{team?.description && (
-					<div>
-						<span className="text-xs font-medium uppercase tracking-wider text-text-muted block mb-1.5">
-							Description
-						</span>
-						<div className="text-[13px] text-text-muted">{team.description}</div>
-					</div>
-				)}
-			</div>
-		</section>
-	);
-}
-
-function AutomationsSection({ teamId }: { teamId: string }) {
-	const { data: team } = useTeam(teamId);
-	const updateTeam = useUpdateTeam(teamId);
-	const wakeOnReply = team?.settings?.wake_mentioner_on_reply ?? true;
-	const savedPageSize = team?.settings?.subtask_page_size ?? DEFAULT_SUBTASK_PAGE_SIZE;
-	const [pageSizeInput, setPageSizeInput] = useState<string>(String(savedPageSize));
-
-	useEffect(() => {
-		setPageSizeInput(String(savedPageSize));
-	}, [savedPageSize]);
-
-	async function toggleWakeOnReply(next: boolean) {
-		await updateTeam.mutateAsync({ settings: { wake_mentioner_on_reply: next } });
-	}
-
-	async function commitPageSize() {
-		const parsed = Math.floor(Number(pageSizeInput));
-		if (!Number.isFinite(parsed) || parsed < 1) {
-			setPageSizeInput(String(savedPageSize));
-			return;
-		}
-		const clamped = Math.min(500, parsed);
-		setPageSizeInput(String(clamped));
-		if (clamped === savedPageSize) return;
-		await updateTeam.mutateAsync({ settings: { subtask_page_size: clamped } });
-	}
-
-	return (
-		<section>
-			<SectionHeader title="Automations" desc="Per-team toggles for agent coordination behavior." />
-			<label className="flex items-start gap-3 text-[13px] max-w-xl cursor-pointer select-none">
-				<input
-					type="checkbox"
-					className="mt-0.5"
-					checked={wakeOnReply}
-					disabled={!team || updateTeam.isPending}
-					onChange={(e) => toggleWakeOnReply(e.target.checked)}
-				/>
-				<span>
-					<span className="font-medium block">Wake mentioner on reply</span>
-					<span className="text-text-muted block mt-0.5">
-						When an agent replies to a comment that @-mentioned it, automatically wake the original
-						commenter so they see the response right away. Disable to have the original commenter
-						pick up accumulated replies on their next heartbeat instead — useful when a single
-						comment @-mentions several agents and you'd rather batch their responses.
-					</span>
-				</span>
-			</label>
-			<div className="mt-5 pt-5 border-t border-border-subtle max-w-xl">
-				<label htmlFor="subtask-page-size" className="block text-[13px] font-medium">
-					Sub-task page size
-				</label>
-				<p className="text-text-muted text-[13px] mt-0.5 mb-1.5">
-					How many sub-tasks to show on an task page before a "Show more" link is offered.
-				</p>
-				<Input
-					id="subtask-page-size"
-					type="number"
-					min={1}
-					max={500}
-					step={1}
-					value={pageSizeInput}
-					onChange={(e) => setPageSizeInput(e.target.value)}
-					onBlur={commitPageSize}
-					onKeyDown={(e) => {
-						if (e.key === 'Enter') (e.target as HTMLInputElement).blur();
-					}}
-					disabled={!team || updateTeam.isPending}
-					className="w-24"
-					data-testid="subtask-page-size-input"
-				/>
-			</div>
-		</section>
-	);
-}
-
-function SecretsSection({ teamId }: { teamId: string }) {
-	const { data: secrets } = useSecrets(teamId);
-	const createSecret = useCreateSecret(teamId);
-	const deleteSecret = useDeleteSecret(teamId);
-	const [showForm, setShowForm] = useState(false);
-	const [name, setName] = useState('');
-	const [value, setValue] = useState('');
-	const [allowedHosts, setAllowedHosts] = useState('');
-	const [allowAllHosts, setAllowAllHosts] = useState(false);
-
-	async function handleCreate(e: React.FormEvent) {
-		e.preventDefault();
-		const hosts = allowedHosts
-			.split(/[\s,]+/)
-			.map((h) => h.trim())
-			.filter(Boolean);
-		if (!allowAllHosts && hosts.length === 0) {
-			alert(
-				'Add at least one allowed host (e.g. api.stripe.com) or check Allow all hosts. ' +
-					'A secret with no hosts cannot be substituted by the egress proxy.',
-			);
-			return;
-		}
-		await createSecret.mutateAsync({
-			name,
-			value,
-			allowed_hosts: hosts,
-			allow_all_hosts: allowAllHosts,
-		});
-		setName('');
-		setValue('');
-		setAllowedHosts('');
-		setAllowAllHosts(false);
-		setShowForm(false);
-	}
-
-	return (
-		<section>
-			<div className="flex items-center justify-between mb-4">
-				<SectionHeader
-					title="Secrets vault"
-					desc="Encrypted secrets agents reference by placeholder. Allowed-hosts gates which upstream hostnames the egress proxy may substitute the secret for — leave empty (and uncheck the override) and the secret is unusable."
-				/>
-				<Button variant="secondary" size="sm" onClick={() => setShowForm(!showForm)}>
-					<Plus className="w-3 h-3" /> Add
-				</Button>
-			</div>
-			{showForm && (
-				<form onSubmit={handleCreate} className="flex flex-col gap-2 mb-3">
-					<div className="flex flex-col sm:flex-row gap-2">
-						<Input
-							placeholder="Name (e.g. STRIPE_API_KEY)"
-							value={name}
-							onChange={(e) => setName(e.target.value)}
-							required
-							className="flex-1"
-						/>
-						<Input
-							placeholder="Value"
-							type="password"
-							value={value}
-							onChange={(e) => setValue(e.target.value)}
-							required
-							className="flex-1"
-						/>
-					</div>
-					<Input
-						placeholder="Allowed hosts (comma- or space-separated, e.g. api.stripe.com, *.stripe.com)"
-						value={allowedHosts}
-						onChange={(e) => setAllowedHosts(e.target.value)}
-						disabled={allowAllHosts}
-					/>
-					<label className="flex items-center gap-2 text-[13px] text-text-muted">
-						<input
-							type="checkbox"
-							checked={allowAllHosts}
-							onChange={(e) => setAllowAllHosts(e.target.checked)}
-						/>
-						Allow this secret to reach any host (escape hatch — use sparingly)
-					</label>
-					<div className="flex gap-2">
-						<Button type="submit" size="sm" disabled={createSecret.isPending}>
-							Add
-						</Button>
-						<Button type="button" variant="secondary" size="sm" onClick={() => setShowForm(false)}>
-							Cancel
-						</Button>
-					</div>
-				</form>
-			)}
-			{secrets?.length === 0 ? (
-				<p className="text-[13px] text-text-subtle">No secrets stored.</p>
-			) : (
-				<div className="flex flex-col gap-1">
-					{secrets?.map((s) => (
-						<div
-							key={s.id}
-							className="flex items-center justify-between rounded-radius-md border border-border bg-bg px-3 py-2 text-[13px]"
-						>
-							<div className="flex items-center gap-2 min-w-0 flex-1">
-								<span className="font-medium">{s.name}</span>
-								<Badge color="neutral">{s.category}</Badge>
-								{s.project_name && (
-									<span className="text-xs text-text-subtle">{s.project_name}</span>
-								)}
-								{s.allow_all_hosts ? (
-									<Badge color="warning">all hosts</Badge>
-								) : s.allowed_hosts?.length ? (
-									<span className="text-xs text-text-subtle font-mono truncate">
-										{s.allowed_hosts.join(', ')}
-									</span>
-								) : (
-									<Badge color="danger">no hosts</Badge>
-								)}
-							</div>
-							<button
-								type="button"
-								onClick={() => deleteSecret.mutate(s.id)}
-								className="text-text-subtle hover:text-accent-red"
-							>
-								<Trash2 className="w-3.5 h-3.5" />
-							</button>
-						</div>
-					))}
-				</div>
-			)}
-		</section>
 	);
 }
